work_1/pdf_to_spreadsheet/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-to-spreadsheet"
version = "1.0.0"
description = "Extrae datos estructurados de PDFs y los exporta a CSV, JSON o Google Sheets"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "PDF to Spreadsheet Team"}
]
keywords = [
    "pdf",
    "extraction",
    "csv",
    "json",
    "google-sheets",
    "data-processing",
    "automation"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
    "Topic :: Text Processing",
    "Topic :: Utilities",
]

dependencies = [
    "pdfplumber>=0.10.0",
    "PyPDF2>=3.0.0",
    "tabula-py>=2.8.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "python-dateutil>=2.8.0",
    "babel>=2.12.0",
    "click>=8.1.0",
    "colorama>=0.4.6",
    "pyyaml>=6.0.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
ocr = [
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "pdf2image>=1.16.0",
]
gsheet = [
    "google-auth>=2.22.0",
    "google-auth-oauthlib>=1.0.0",
    "gspread>=5.10.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "flake8>=6.1.0",
    "black>=23.7.0",
    "mypy>=1.5.0",
    "isort>=5.12.0",
]
all = [
    "pdf-to-spreadsheet[ocr,gsheet,dev]",
]

[project.scripts]
pdf2sheet = "main:main"

[project.urls]
Homepage = "https://github.com/tu-usuario/pdf-to-spreadsheet"
Documentation = "https://github.com/tu-usuario/pdf-to-spreadsheet#readme"
Repository = "https://github.com/tu-usuario/pdf-to-spreadsheet.git"
Issues = "https://github.com/tu-usuario/pdf-to-spreadsheet/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
